requirements.txt
aiosmtplib==2.0.1
alembic==1.10.4
anyio==3.6.2
bcrypt==4.0.1
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
encryptoenv==0.0.2
exceptiongroup==1.1.1
fastapi==0.95.1
fastapi-mail==1.2.8
fileflamingo==0.0.6
filelock==3.12.0
flake8==6.0.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mirakuru==2.5.1
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
port-for==0.6.3
pre-commit==3.3.1
psutil==5.9.5
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.10.1
pydantic==1.10.7
pyflakes==3.0.1
pytest==7.3.1
pytest-postgresql==4.1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.30.0
rsa==4.9
ruff==0.0.263
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
virtualenv==20.23.0
